♻️ Moved Zabbix-to-file methods into respective handlers
